+++ bundler-packages/tevm-run/src/tevm-run.js
@@ -1,0 +1,10 @@
+#!/usr/bin/env bun
+
+import('./run.js')
+	.then(({ run }) => {
+		return run()
+	})
+	.catch((e) => {
+		console.error('An error occurred:', e)
+		process.exit(1)
+	})