+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tevm/server
 
+## 2.0.0-next.103
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tevm/memory-client@2.0.0-next.103
+  - @tevm/decorators@2.0.0-next.103
+
 ## 2.0.0-next.102
 
 ### Patch Changes